feat: move login logic into seperated service (#4)
https://grove-sandwich-1dd.notion.site/Views-ViewModels-Create-Login-Register-pages-d52bdfe4be5349418c625bdcc2144e52?pvs=4
